backend/requirements.txt
boto3==1.26.68
Django==4.1
djangorestframework==3.13.1
djoser==2.1.0
requests==2.28.1
drf_nested_routers==0.93.4
psycopg2==2.9.3
django-filter==22.1
django-import-export==2.8.0
django-storages==1.13.2
django-templated-mail==1.1.1
uWSGI==2.0.24
celery==5.2.7
redis==4.3.4
flower==1.2.0
django-cors-headers==3.13.0
more-itertools==8.14.0
pillow==9.2.0
drf-yasg==1.21.4
pytest==7.2.0
drf-oidc-auth==3.0.0
pytest-django==4.5.2
model-bakery==1.9.0
mozilla-django-oidc==3.0.0
sentry-sdk==1.39.1
django-health-check==3.18.1
pydantic==2.7.4